+++ response/pagination.go
@@ -2,9 +2,9 @@ package response
 
 // PaginationResponse represents pagination details
 type Pagination struct {
-	PageNumber int `json:"pageNumber"`
-	PageSize   int `json:"pageSize"`
-	Total      int `json:"total"`
+	PageNumber int   `json:"pageNumber"`
+	PageSize   int   `json:"pageSize"`
+	Total      int64 `json:"total"`
 }
 
 // RestResponsePagination represents a paginated response
